+++ samples/applets/share-document.js
@@ -4,8 +4,6 @@
  * <cliend_id>, <client_secret>, <username>, <password>, <document_id> - are required params
  */
 
-// node samples/applets/share-document 6fa29da4a3308e55c63724eaf30f1cdc ed811208b2594d9e0564df6313f423fe zolotov.artem@pdffiller.team 12345678 3eee0e56d6e62fa2b96e384a79e7ab08e4762201
-
 'use strict';
 
 const [
@@ -13,7 +11,7 @@ const [
   clientSecret,
   username,
   password,
-  document_id,
+  documentId,
 ] = process.argv.slice(2);
 
 const api = require('../../lib')({
@@ -34,7 +32,7 @@ getAccessToken({
     const { access_token: token } = tokenRes;
 
     shareDocument({
-      id: document_id,
+      id: documentId,
       token,
     }, (shareErr, shareRes) => {
       if (shareErr) {