+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Created [Applet](https://github.com/signnow/SignNowNodeSDK/blob/master/samples/applets/share-document.js) for testing of creation download link with *Share Document* method
-- Documented *Share Document* method in [*document*](https://signnow.github.io/SignNowNodeSDK/class/lib/document.js~Document.html) class
+- Documented *Share Document* feature in [*Document*](https://signnow.github.io/SignNowNodeSDK/class/lib/document.js~Document.html) class
 
 ### Changed